requirements.txt
environs>=14.3.0
logger>=1.4
python-telegram-bot==13.0
requests>=2.32.4